+++ src/main/java/net/daw/dao/implementation/UsuarioDao.java
@@ -59,9 +59,9 @@ public class UsuarioDao implements ViewDaoInterface<UsuarioBean>, TableDaoInterf
     }
 
     @Override
-    public int getCount(ArrayList<FilterBeanHelper> hmFilter) throws Exception {
+    public Long getCount(ArrayList<FilterBeanHelper> hmFilter) throws Exception {
         strSQL += SqlBuilder.buildSqlWhere(hmFilter);
-        int pages = 0;
+        Long pages = 0L;
         try {
             pages = oMysql.getCount(strSQL);
         } catch (Exception ex) {
